requirements.txt
opencv-python
cmake
dlib
tensorflow-gpu
scipy
numpy
Pandas
tqdm
h5py
hydra-core
scikit-learn
albumentations
omegaconf
wandb
